Colocate all stylesheets and graphql queries
