+++ src/pages/index.tsx
@@ -1,6 +1,6 @@
 import { graphql } from 'gatsby'
 
-import { withLayout } from '../components/layout-wrapper';
+import { withLayout } from '../components/application-wrapper';
 import { IndexPage } from '../components/static-pages/index';
 
 // graphql fragments eventFields
