+++ src/stylesheets/_footer.scss
@@ -1,4 +1,4 @@
-@import '../../colors';
+@import 'colors';
 
 .footer {
   color: $theme-light-gray;
