+++ src/stylesheets/application.scss
@@ -61,9 +61,8 @@
 // @import "node_modules/bootstrap/scss/carousel";
 @import "node_modules/bootstrap/scss/utilities";
 
-@import "../colors";
-
-@import "footer/footer";
+@import "colors";
+@import "footer";
 
 // page-level layout
