+++ src/templates/article.tsx
@@ -1,9 +1,18 @@
 import { graphql } from 'gatsby'
 
-import { withLayout } from '../components/layout-wrapper';
+import { withLayout } from '../components/application-wrapper';
 import { ArticleTemplate } from '../components/article';
 
 export default withLayout(ArticleTemplate)
+
+export const smappExportFields = graphql`
+fragment smappExportFields on SmappExportCsv {
+  studentName
+  amount
+  designation
+  name
+}
+`
 
 export const pageQuery = graphql`
   query ArticleByPath($path: String!) {
